Make health status work
Calling .jobs on CookbookWorker was always resulting in a no method
error and wasn't actually trying to communicate with redis, so if redis
wasn't around it wouldn't throw an exception. This changes it to use the
sidekiq API instead which will raise a redis exception if redis isn't up
and running. Also adds some additional points of data to the status
endpoint.